保存処理を TimelineConverter クラスに持っていく
